+++ lib/pengine/clone.c
@@ -570,7 +570,7 @@ clone_print(resource_t * rsc, const char *pre_text, long options, void *print_da
 int
 pe__clone_xml(pcmk__output_t *out, va_list args)
 {
-    long options = va_arg(args, long);
+    unsigned int options = va_arg(args, unsigned int);
     pe_resource_t *rsc = va_arg(args, pe_resource_t *);
 
     GListPtr gIter = rsc->children;
@@ -598,7 +598,7 @@ pe__clone_xml(pcmk__output_t *out, va_list args)
 int
 pe__clone_html(pcmk__output_t *out, va_list args)
 {
-    long options = va_arg(args, long);
+    unsigned int options = va_arg(args, unsigned int);
     pe_resource_t *rsc = va_arg(args, pe_resource_t *);
 
     char *list_text = NULL;
@@ -785,7 +785,7 @@ pe__clone_html(pcmk__output_t *out, va_list args)
 int
 pe__clone_text(pcmk__output_t *out, va_list args)
 {
-    long options = va_arg(args, long);
+    unsigned int options = va_arg(args, unsigned int);
     pe_resource_t *rsc = va_arg(args, pe_resource_t *);
 
     char *list_text = NULL;
