+++ lib/pengine/native.c
@@ -1072,7 +1072,7 @@ native_print(resource_t * rsc, const char *pre_text, long options, void *print_d
 int
 pe__resource_xml(pcmk__output_t *out, va_list args)
 {
-    long options = va_arg(args, int);
+    unsigned int options = va_arg(args, unsigned int);
     pe_resource_t *rsc = va_arg(args, pe_resource_t *);
 
     const char *class = crm_element_value(rsc->xml, XML_AGENT_ATTR_CLASS);
@@ -1139,7 +1139,7 @@ pe__resource_xml(pcmk__output_t *out, va_list args)
 int
 pe__resource_html(pcmk__output_t *out, va_list args)
 {
-    long options = va_arg(args, int);
+    unsigned int options = va_arg(args, unsigned int);
     pe_resource_t *rsc = va_arg(args, pe_resource_t *);
     node_t *node = pe__current_node(rsc);
 
@@ -1156,7 +1156,7 @@ pe__resource_html(pcmk__output_t *out, va_list args)
 int
 pe__resource_text(pcmk__output_t *out, va_list args)
 {
-    long options = va_arg(args, int);
+    unsigned int options = va_arg(args, unsigned int);
     pe_resource_t *rsc = va_arg(args, pe_resource_t *);
 
     node_t *node = pe__current_node(rsc);
